Code changes to stay up to date with Laravel 13 [API-464]

+++ database/migrations/2026_09_02_115707_create_cache_locks_table.php
@@ -13,7 +13,7 @@ return new class () extends Migration {
         Schema::create('cache_locks', function (Blueprint $table) {
             $table->string('key')->primary();
             $table->string('owner');
-            $table->integer('expiration')->index();
+            $table->bigInteger('expiration')->index();
         });
     }
